Renamed boolean matrix
