+++ project/__init__.py
@@ -1,8 +1,8 @@
 import project.automata_utils
 import project.graph_utils
-import project.matrices_utils
+import project.boolean_matrix
 import project.rpq
 from project.automata_utils import *
 from project.graph_utils import *
-from project.matrices_utils import *
+from project.boolean_matrix import *
 from project.rpq import *
